fix: [config] add explicit relative path field to the config loading
The rel_path is a calculated field that is critical to determining where
files are placed and found in the scratch dir and also the reports dir.
At the moment the field is not created until the flow objects are
created. Which means that the saved config doesn't have the information
unless it's overridden in the config itself.

This commit moves the calculated field out of the flow base class init
into the config loading stage where it is saved the filesystem in the
scratch dir along with the rest of the loaded config. Which means it can
be used to navigate the scratch dir after a run.

Signed-off-by: James McCorrie <james.mccorrie@lowrisc.org>