refactor: [config] move overrides from flow base to config loading stage
This is to ensure that the saved project config is already resolved
otherwise the overrides mechanism would need to be implemented in
any code that loads the project config.

Without this change, the report generation code can't see the correct
rel_path field as it is overridden in the flow object after being saved
as part of the project config.

Signed-off-by: James McCorrie <james.mccorrie@lowrisc.org>