style: disable TRY003
Disable rule that wants long exception error messages to be defined in
the Error class itself. This makes some sense for custom error classes.
However it's not appropriate to define custom error classes for every
different error case. Generally a common error class is used in such
cases, for example `RuntimeError` and `ValueError`.

Signed-off-by: James McCorrie <james.mccorrie@lowrisc.org>