fix: remove shebangs from scripts
It doesn't look like these scripts are used in this version of DVSim.
Instead the versions shipped with OpenTitan are used if they are used at all.

For the moment remove the shebangs and remove the executable flag from
the files. They would not be useable from a package install anyway. When
it comes to reintegrating with OpenTitan, then a new solution will be
required. For the moment this patch resolves new failing linting checks.

Signed-off-by: James McCorrie <james.mccorrie@lowrisc.org>