fix: circular import issue
When adding typing I've introduced a circular import issue. This
generally highlights a bigger structural issue, which should be
resolved. For the moment I've added a type check guard around the import
as a short term fix.

Longer term we should find a better way of catching issues like this.

Signed-off-by: James McCorrie <james.mccorrie@lowrisc.org>