refactor: use deployment name instead of object as dict keys
Update the scheduler to index the deployment object mappings by
`deploy.full_name` instead of the `deploy` object itself. This means
that the `Deploy` object does not need to be "hash-able".

The `full_name` now takes account of the variant, as without this the
deployment objects are not uniquely identifiable from the "full name".
From what I can tell the only use of the `full_name` attribute is for
logs and for unique IDs, which turns out not to be unique enough.

Signed-off-by: James McCorrie <james.mccorrie@lowrisc.org>