test: add initial CompileSim unittest
Add a `new()` static method to allow tests to test against the desired
interface. The intention is to convert the deployment classes to
Pydantic typed data classes. The deployment classes seem to be builders
of fairly simple data classes once the construction process has been
completed. The constructors take the data in other formats and then
filter and transform to result in a new data class (deployment object).

Pydantic data class constructors take in the final data, so we need a
function to take the original data and do the transform. Having a
`new()` static method is a fairly standard convention in languages such
as Rust. For the moment this method just delegates to the existing
constructor. The `CompileSim` unittests make use of this new method
instead of the constructor directly. Which means they will be able to
test the new code as well as the old code for A/B testing, without
having to continually modify the tests.

Signed-off-by: James McCorrie <james.mccorrie@lowrisc.org>