feat: add ResultsSummary model
At the moment we have a JSON file for each of the flow config objects,
alongside the HTML reports. The results summary page which ties them all
together is HTML only. There is no way of getting the information
programmatically, without doing something like parsing the HTML... which
is error prone and messy.

The ResultsSummary is a Pydantic model which can then be dumped to JSON.

Signed-off-by: James McCorrie <james.mccorrie@lowrisc.org>