feat: summary report more dashboard like
Introduce a CoverageMetrics model instead of the dict. Create a new
method on the tool plugins to convert from eda tool specific coverage
fields to generic names that match the opentitan-site dashboard fields.

Update the report summary and block report templates to use the new
model. Add the extra fields in the summary report to make it more like
the dashboard (contain the same information).

Signed-off-by: James McCorrie <james.mccorrie@lowrisc.org>