feat: add git utils for getting git commit hash
The current method for getting the git commit hash is to evaluate git
commands as part of hjson wildcard substitution. Presumably as a means
of making it more generic and supporting other version control systems.

However the opentitan HJSON builds up a markdown string containing a URL
link to github. Then DVSim parses that string with a regex. Which
couples the somewhat generic mechanism to an exact implementation.

This commit adds some generic git utility functions to work directly
with git from within DVSim and bypass HJSON calculated values. If we are
going to be calling Git, I'd rather do that directly from DVSim, rather
than some uncontrolled system call eval in HJSON. If we need to add
flexibility to support different systems then we can add that in the
future using a plugin system in DVSim and declaratively switching
between version control system (i.e. git) and hosting solution (i.e.
github) in the HJSON if required.

The current solution is highly coupled and doesn't allow for extension.
This solution is still git/github specific, but adds scope for later
extension if required. It also removes the dependence on one more eval
in the OpenTitan HJSON.

Signed-off-by: James McCorrie <james.mccorrie@lowrisc.org>